2024年5月13日
aw30n-release_V1.2.0
SDK 等级： B
发布时间： 2024 年 5 月 13 日
发布版本： aw30n-release_V1.2.0
发布说明：
1、 新增 audio_link 驱动；
2、 新增全双工对讲机应用；
3、 新增实时调节蓝牙发射功率 API；
4、 新增 adpcm 编码输出 header 功能；
5、 优化蓝牙底层库；
6、 优化从机手机 OTA 升级兼容性；
7、 优化 gpio、 usb、 uart、 io_wakeup 驱动；
8、 优化不可屏蔽中断流程；
9、 优化线性移频算法；
10、 修复测试盒串口升级失败问题；
11、 修复 saradc dm/dp 采集错误问题；
12、 修复 MCPWM 注销时异常和固定 IO 输出 bug；
13、 修复软关机流程获取 key 错误导致软关机复位源信息错误问题；
文档更新：
1、 更新 SDK 文档《AW30N_SDK 手册_V1.4.pdf》；
2、 更新芯片手册《AW30N_芯片手册_V1.1.pdf》；
3、 新增硬件规格书《AW302C Datasheet V1.0.pdf》
4、 更新硬件规格书《AW302A Datasheet V1.1.pdf》、 《AW305A Datasheet V1.3.pdf》、
《AW306A Datasheet V1.4.pdf》；
5、 新增硬件原理图资料；
工具版本：
1、 一拖二烧写器： v2.27.8， 通过软件上远程更新
2、 一拖八烧写器： v3.1.23， 通过软件上远程更新
3、 蓝牙测试盒： v1.3.2， 通过软件上远程更新
4、 蓝牙 profile 制作工具 make_gatt_services： v1.2.1， 通过 gitee 下载
5、 打包工具 & 音频文件转换工具通过 gitee 下载： 打包工具&音频文件转换工具
6、 工具链： 2.4.9
7、 密钥版本： 派生密钥 2.0

+++ sdk/apps/app/bsp/common/update/dev_update.c
@@ -399,7 +399,7 @@ u16 dev_update_check(char *logo)
         }
 
         FS_PARTITION_INFO part_info = {0};
-        res = vfs_ioctl(pfs, FS_IOCTL_GET_PARTITION_INFO, &part_info);
+        res = vfs_ioctl(pfs, (int)FS_IOCTL_GET_PARTITION_INFO, (int)&part_info);
         if (res) {
             log_info(">>>[test]:dev part err!!!\n");
             fs_fs_close(&pfs);
